modify: refactoring
