+++ src/pages/hansolKim/Main/Aside/Aside.js
@@ -1,20 +1,21 @@
 import React from 'react';
 import Story from './Story';
-import Recommen from './Recommen';
+import Recommend from './Recommend';
 import Username from './Username';
+import './Aside.scss';
 
-const Content3 = () => {
+const Aside = () => {
   return (
-    <>
+    <div className="aside">
       <Username />
       <Story />
-      <Recommen />
-      <div className="copyright text2">
+      <Recommend />
+      <div className="copyright normalText">
         Westagram 정보 . 지원 . 홍보센터 . API . 채용 정보 . 개인정보처리방침 .
         약관 . 디렉터리 . 프로필 . 해시태그 . 언어 {'\n'} {'\n'}ⓒ 2023 WESTAGRAM
       </div>
-    </>
+    </div>
   );
 };
 
-export default Content3;
+export default Aside;
