+++ src/pages/hansolKim/Main/Aside/Aside.scss
@@ -1,0 +1,12 @@
+.aside {
+    display: flex;
+    flex-direction: column;
+    width: 250px;
+    background-color: #fafafa;
+    .copyright {
+      margin-top: 10px;
+      white-space: pre-wrap;
+      font-size: 12px;
+      color: #9b9b9b;
+    }
+  }
