+++ src/pages/hansolKim/Main/Aside/CopyRight.js
@@ -1,0 +1,14 @@
+export const Copy_Right_List = [
+  { id: 1, listName: Westagram },
+  { id: 2, listName: Westagram },
+  { id: 3, listName: Westagram },
+  { id: 4, listName: Westagram },
+  { id: 5, listName: Westagram },
+  { id: 6, listName: Westagram },
+  { id: 7, listName: Westagram },
+  { id: 8, listName: Westagram },
+  { id: 9, listName: Westagram },
+  { id: 10, listName: Westagram },
+  { id: 11, listName: Westagram },
+  { id: 12, listName: Westagram },
+];
