+++ src/pages/hansolKim/Main/Aside/Recommend.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import './Recommend.scss';
+
+const Recommend = () => {
+  return (
+    <div className="recommend">
+      <div className="recommendLine">
+        <span className="recommendText">회원님을 위한 추천</span>
+        <span className="recommendRightText">모두 보기</span>
+      </div>
+    </div>
+  );
+};
+
+export default Recommend;
