+++ src/pages/hansolKim/Main/Aside/Username.js
@@ -1,12 +1,13 @@
 import React from 'react';
+import './Username.scss';
 
 const Username = () => {
   return (
     <div className="username">
       <img src="images/hansolKim/profile.png" alt="profile" />
       <div className="textLine">
-        <div className="text1">noname123</div>
-        <div className="text2">noname | 무제</div>
+        <div className="boldText">noname123</div>
+        <div className="userText">noname | 무제</div>
       </div>
     </div>
   );
