+++ src/pages/hansolKim/Main/Aside/Username.scss
@@ -1,0 +1,23 @@
+.username {
+    display: flex;
+    align-items: center;
+    width: 250px;
+    height: 50px;
+    margin-top: 100px;
+    background-color: #fafafa;
+    img {
+      width: 40px;
+      height: 40px;
+      padding: 5px;
+    }
+    .textLine {
+      font-size: 12px;
+      .boldText {
+        font-size: 12px;
+        font-weight: bold;
+      }
+      .userText {
+        font-size: 12px;
+      }
+    }
+  }
