--- src/pages/hansolKim/Main/Content1/Content1.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import Navbar from './Navbar';
-
-const Content1 = () => {
-  return <Navbar />;
-};
-
-export default Content1;
