--- src/pages/hansolKim/Main/Content3/Recommen.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Recommen = () => {
-  return (
-    <div className="recommend">
-      <div className="recommendLine">
-        <span className="text2">회원님을 위한 추천</span>
-        <span className="text3">모두 보기</span>
-      </div>
-    </div>
-  );
-};
-
-export default Recommen;
