+++ src/pages/hansolKim/Main/Main.js
@@ -1,23 +1,17 @@
 import React from 'react';
 import './Main.scss';
-import Content1 from './Content1/Content1';
-import Content2 from './Content2/Content2';
-import Content3 from './Content3/Content3';
+import Nav from './Nav/Nav';
+import Article from './Article/Article';
+import Aside from './Aside/Aside';
 
 const Main = () => {
   return (
-    <section>
+    <section className="main">
       <div className="layout">
-        <div className="content1">
-          <Content1 />
-        </div>
-        <div className="main">
-          <div className="content2">
-            <Content2 />
-          </div>
-          <div className="content3">
-            <Content3 />
-          </div>
+        <Nav />
+        <div className="mainArea">
+          <Article />
+          <Aside />
         </div>
       </div>
     </section>
