+++ src/pages/hansolKim/Main/Main.scss
@@ -1,192 +1,13 @@
-@import '../../../styles/variables.scss';
-
-$backgroundColor: #ffffff;
-img {
-  width: 20px;
-  height: 20px;
-}
-.text1 {
-  font-size: 12px;
-  font-weight: bold;
-}
-.text2 {
-  font-size: 12px;
-  color: #9b9b9b;
-}
-.text3 {
-  font-size: 12px;
-}
-
 .layout {
   width: 100%;
-  height: 100vh;
+  height: 100%;
   margin: 0;
   background-color: #fafafa;
   overflow-y: visible;
-  .content1 {
-    @include flex(space-evenly, center);
-    position: fixed;
-    background-color: #ffffff;
-    padding: 10px;
-    width: 100%;
-    border-bottom: 1px solid #dbdbdb;
-    .navbar-left {
-      font-family: 'Lobster';
-      font-size: 25px;
-    }
-    .navbar-center {
-      .search {
-        width: 230px;
-        padding: 5px;
-        background-color: #fafafa;
-        border: 1px solid #dbdbdb;
-        text-align: center;
-      }
-    }
-    .navbar-right {
-      img {
-        margin: 10px;
-      }
-    }
-  }
-  .main {
+  .mainArea {
     display: flex;
     justify-content: center;
     min-height: 75vh;
     background-color: #fafafa;
-    .content2 {
-      width: 500px;
-      padding-left: 30px;
-      background-color: #fafafa;
-      .article {
-        display: flex;
-        flex-direction: column;
-        width: 450px;
-        min-height: 300px;
-        margin-top: 100px;
-        background-color: #ffffff;
-        border: 1px solid #dbdbdb;
-        .userLine {
-          display: flex;
-          justify-content: space-between;
-          padding: 10px;
-          .userLine-left {
-            display: flex;
-            align-items: center;
-            padding-left: 5px;
-            span {
-              padding-left: 5px;
-              font-weight: bold;
-              font-size: 12px;
-            }
-          }
-        }
-        .imageLine {
-          img {
-            width: 450px;
-            height: 450px;
-          }
-        }
-        .imojiLine {
-          display: flex;
-          justify-content: space-between;
-          padding: 5px;
-          .imoji-left {
-            padding: 5px;
-            img {
-              width: 18px;
-              height: 18px;
-              margin-right: 5px;
-            }
-          }
-        }
-        .likeLine {
-          display: flex;
-          align-items: center;
-          padding-left: 5px;
-          span {
-            padding-left: 3px;
-            font-weight: bold;
-            font-size: 12px;
-          }
-        }
-        .commentBox {
-          min-height: 80px;
-          font-size: 12px;
-          font-weight: bold;
-          padding-left: 10px;
-        }
-        .commentLine {
-          position: relative;
-          input {
-            width: 100%;
-            padding: 10px;
-            border: 1px solid #dbdbdb;
-          }
-          .postBtn {
-            position: absolute;
-            top: 15px;
-            right: 15px;
-            border: none;
-            background-color: #ffffff;
-            color: #0095f6;
-            cursor: pointer;
-            &:disabled {
-              color: #c4e1fb;
-              cursor: default;
-            }
-          }
-        }
-      }
-    }
-    .content3 {
-      display: flex;
-      flex-direction: column;
-      width: 250px;
-      background-color: #fafafa;
-      .username {
-        display: flex;
-        align-items: center;
-        width: 250px;
-        height: 50px;
-        margin-top: 100px;
-        background-color: #fafafa;
-        img {
-          width: 40px;
-          height: 40px;
-          padding: 5px;
-        }
-        .textLine {
-          font-size: 12px;
-        }
-      }
-      .story {
-        width: 250px;
-        height: 150px;
-        background-color: #ffffff;
-        border: 1px solid #dbdbdb;
-        .storyLine {
-          display: flex;
-          justify-content: space-between;
-          padding: 10px;
-        }
-      }
-      .recommend {
-        width: 250px;
-        height: 120px;
-        margin-top: 20px;
-        background-color: #ffffff;
-        border: 1px solid #dbdbdb;
-        .recommendLine {
-          display: flex;
-          justify-content: space-between;
-          padding: 10px;
-        }
-      }
-      .copyright {
-        margin-top: 10px;
-        white-space: pre-wrap;
-      }
-    }
   }
 }
